chore: remove platform-sdk file from polkadot module
