+++ libs/ledger-live-common/src/families/polkadot/platformAdapter.test.ts
