+++ libs/ledger-live-common/src/families/polkadot/platformAdapter.ts
